feat: update career schema and adjust flight table constraints; refactor tools to use new model
