+++ apps/mcp-health/src/tools/mental-health.ts
@@ -44,7 +44,7 @@ export function registerMentalWellnessTool(server: McpServer) {
         })
 
         const response = await generateObject({
-          model: lmstudio('qwen2.5-7b-instruct-1m'),
+          model: lmstudio('gemma-3-12b-it'),
           prompt,
           schema: MentalWellnessSchema,
         })
