+++ apps/mcp-health/src/tools/sleep.ts
@@ -44,7 +44,7 @@ export function registerSleepTool(server: McpServer) {
         })
 
         const response = await generateObject({
-          model: lmstudio('qwen2.5-7b-instruct-1m'),
+          model: lmstudio('gemma-3-12b-it'),
           prompt,
           schema: SleepAnalysisSchema,
         })
