+++ apps/mcp-health/src/tools/workout.ts
@@ -47,7 +47,7 @@ export function registerWorkoutTool(server: McpServer) {
         })
 
         const response = await generateObject({
-          model: lmstudio('qwen2.5-7b-instruct-1m'),
+          model: lmstudio('gemma-3-12b-it'),
           prompt,
           schema: WorkoutRecommendationSchema,
         })
