+++ apps/web/components/career/job-application.form.tsx
@@ -1,7 +1,8 @@
+import { JobApplicationStage } from '@/lib/career'
 import { useCreateApplication, useDeleteApplication } from '@/lib/hooks/useApplications'
 import { useAuth } from '@clerk/nextjs'
 import type { JobApplication } from '@ponti/utils/career'
-import { JobApplicationStage, JobApplicationStatus } from '@ponti/utils/types'
+import { JobApplicationStatus } from '@ponti/utils/types'
 import { Plus, PlusCircle } from 'lucide-react'
 import { useCallback, useMemo, useState, type FormEvent } from 'react'
 import { CompanySelect } from '../company-select'
