+++ packages/utils/src/career/job.service.ts
@@ -1,8 +1,8 @@
 import { eq, type SQL } from 'drizzle-orm'
 import { createInsertSchema, createSelectSchema } from 'drizzle-zod'
 import { db } from '../db/index'
+import { jobs, type Job, type NewJob } from '../db/schema/career.schema'
 import { companies } from '../db/schema/company.schema'
-import { jobs, type Job, type NewJob } from '../db/schema/job.schema'
 
 export const JobInsertSchema = createInsertSchema(jobs)
 export const JobSchema = createSelectSchema(jobs)
