+++ packages/utils/src/db/0008_wise_iron_patriot.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "flight" ALTER COLUMN "reservationNumber" DROP NOT NULL;--> statement-breakpoint
+ALTER TABLE "flight" ALTER COLUMN "url" DROP NOT NULL;
