+++ packages/utils/src/db/meta/_journal.json
@@ -57,6 +57,13 @@
       "when": 1742341100095,
       "tag": "0007_conscious_shinobi_shaw",
       "breakpoints": true
+    },
+    {
+      "idx": 8,
+      "version": "7",
+      "when": 1742366209189,
+      "tag": "0008_wise_iron_patriot",
+      "breakpoints": true
     }
   ]
 }
