+++ packages/utils/src/db/schema/index.ts
@@ -2,13 +2,13 @@ export * from './activity.schema'
 export * from './auth.schema'
 export * from './bookmarks.schema'
 export * from './calendar.schema'
+export * from './career.schema'
 export * from './chats.schema'
 export * from './company.schema'
 export * from './documents.schema'
 export * from './finance.schema'
 export * from './health.schema'
 export * from './items.schema'
-export * from './job.schema'
 export * from './lists.schema'
 export * from './movies.schema'
 export * from './music.schema'
