+++ packages/utils/src/db/schema/travel.schema.ts
@@ -13,8 +13,8 @@ export const flight = pgTable(
     arrivalDate: timestamp('arrivalDate', { mode: 'string' }).notNull(),
     arrivalAirport: text('arrivalAirport').notNull(),
     airline: text('airline').notNull(),
-    reservationNumber: text('reservationNumber').notNull(),
-    url: text('url').notNull(),
+    reservationNumber: text('reservationNumber'),
+    url: text('url'),
     createdAt: timestamp('createdAt', { precision: 3, mode: 'string' }).defaultNow().notNull(),
     updatedAt: timestamp('updatedAt', { precision: 3, mode: 'string' }).defaultNow().notNull(),
     userId: uuid('userId').notNull(),
