+++ packages/utils/src/types/index.ts
@@ -1,4 +1,4 @@
 export * from '../db/schema/activity.schema'
-export * from '../db/schema/job.schema'
+export * from '../db/schema/career.schema'
 export * from '../db/schema/music.schema'
 export * from './health-and-fitness'
